\name{ISspectrum-package}
\alias{ISspectrum-package}
\alias{ISspectrum}
\docType{package}
\title{
Incoherent scatter spectrum calculation
}
\description{
R-interface to GUISDAP C-routines for incoherent scatter spectrum calculation
}
\details{
\tabular{ll}{
Package: \tab ISspectrum\cr
Version: \tab 0.1-1\cr
Date: \tab 2013-06-04\cr
License: \tab \cr
Depends: \tab \cr
}

Index:

\code{\link{ISspectrum.general}} Incoherent scatter spectrum with free
selection of parameters for each ion species
\code{\link{ISspectrum.guisdap}} GUISDAP -style incoherent scatter
spectrum with equal temperature, speed, and collision frequency of all
ion species 
}
\author{
Ilkka Virtanen (University of Oulu, Finland) \cr \email{ilkka.i.virtanen@oulu.fi}
}
\keyword{ package }
