\name{ISspectrum-package}
\alias{ISspectrum-package}
\alias{ISspectrum}
\docType{package}
\title{
Incoherent scatter spectrum calculation
}
\description{
R-interface to GUISDAP C-routines for incoherent scatter spectrum calculation
}
\details{
\tabular{ll}{
Package: \tab ISspectrum\cr
Version: \tab 0.1-0\cr
Date: \tab 2011-01-11\cr
License: \tab \cr
Depends: \tab \cr
Built: \tab R 2.14.1; universal-apple-darwin9.8.0; 2012-01-10 10:26:17 UTC; unix\cr
}

Index:

\code{\link{ISspectrum.general}} Incoherent scatter spectrum with free
selection of parameters for each ion species
\code{\link{ISspectrum.guisdap}} GUISDAP -style incoherent scatter
spectrum with equal temperature, speed, and collision frequency of all
ion species 
}
\author{
Ilkka Virtanen (University of Oulu, Finland) \cr \email{ilkka.i.virtanen@oulu.fi}
}
\keyword{ package }
